move test to unit tests for deps
